pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "strands-mlx"
version = "0.2.4"
description = "Use MLX in Strands Agents"
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = {text = "MIT"}
authors = [
    {name = "Cagatay Cali", email = "cagataycali@icloud.com"}
]
keywords = [
    "strands",
    "agents",
    "fine-tuning",
    "lora",
    "llm",
    "mlx-lm",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "strands-agents",
    "huggingface_hub",
    "mlx-lm",
    "mlx-vlm"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio",
    "black",
    "isort",
    "mypy",
    "ruff",
]
vision = [
    "mlx-vlm>=0.3.0",
    "pillow>=10.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "soundfile>=0.13.0",
]
all = [
    "strands-agents-tools",
    "pytest>=7.0",
    "pytest-asyncio",
    "black",
    "isort",
    "mypy",
    "ruff",
    "mlx-vlm>=0.3.0",
    "pillow>=10.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "soundfile>=0.13.0"
]

[project.urls]
Homepage = "https://github.com/cagataycali/strands-mlx"
Repository = "https://github.com/cagataycali/strands-mlx"
Issues = "https://github.com/cagataycali/strands-mlx/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["strands_mlx*"]
exclude = ["tests*", "test_media*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
